Added parsing JSON with serde_json
I have added parsing to the new() function that parses the input jsons
into serde_json::Values. I have also updated the tests for this and have
modified the build process to allow for building on nightly for
debugging purposes (better error output) and for building on stable for
release.